update #9: Split EntityInterface into more solid interfaces
